tests: fix iq2xxs test harnesses and add MUL_MAT_ID coverage
Harness fixes (library code untouched):
- weight generator: ru()%200-100 underflowed in uint64 domain, turning
  ~50% of scales into fp16 +Inf; NaN comparisons made the old 'bitexact'
  result vacuous. Use (int)(ru()%200)-100 and count NaNs in comparisons.
- kernel test: repack8 call passed 8 consecutive blocks of one row where
  real repack gathers the same-index block of 8 rows (src[x + i*nblocks]).
- dispatch test: MUL_MAT comparison index was transposed (dst is ne[0]-
  contiguous): r2[t*m + row], not r2[row*n_tokens + t].

With valid data: gemv/gemm generic vs vec_dot ref truly bitexact
(0.000e+00), avx512 vs generic 4.7e-05 (FP reorder noise).

New test-iq2xxs-mmid: exercises repack IQ2_XXS MUL_MAT_ID in DSV4 expert
shapes (k=2048 m=512 n_as=32 used=8) at decode (tok=1) and pp (tok=64);
max noise 7.6e-07, nan=0. PP 2.16x faster than unoptimized path; decode
0.64x slower (small-batch scheduling overhead, not a kernel issue).

+++ test-iq2xxs-bench.cpp
@@ -30,7 +30,7 @@ static bench_result bench_mmid(ggml_backend_t backend, int k, int m, int n_mats,
     for (auto & b : w) {
         uint64_t * p = (uint64_t *) b.qs;
         for (size_t i = 0; i < sizeof(b.qs) / 8; i++) p[i] = ru();
-        b.d = ggml_fp32_to_fp16(0.01f * (float) (ru() % 200 - 100));
+        b.d = ggml_fp32_to_fp16(0.01f * (float) ((int) (ru() % 200) - 100));
     }
     std::vector<float> act((size_t) k * n_tokens);
     for (auto & v : act) v = 0.5f * ((float) (ru() % 2000) / 1000.0f - 1.0f);
